+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- New prop `forbiddenRedirectPath`.
+
 ## [0.1.1] - 2019-11-06
 
 ### Fixed